无法识别性别, 生成的物品没有持久

+++ DB/src/main/java/com/zhaoxiaodan/mirserver/db/entities/PlayerItem.java
@@ -20,7 +20,7 @@ public class PlayerItem implements Parcelable{
 	@Embedded
 	public ItemAttr attr;
 
-	public short dura;
+	public int dura;
 
 	public boolean isWearing;
